requirements.txt
Django==3.2
celery==5.0.5
gunicorn==20.0.4
django-extensions==3.1.1
django-filter==2.4.0
django-jazzmin==2.4.4
djangorestframework==3.12.2
firebase-admin==4.5.2
Pillow==8.1.0
python-decouple==3.4
redis==3.5.3
psycopg2-binary==2.8.6
djangorestframework-simplejwt==4.6.0
django-summernote==0.8.11.6
django-cors-headers==3.7.0
django-cleanup==5.1.0
coreapi==2.3.3